fix: refactored types and updated axios lib vunerability
