+++ package.json
@@ -25,7 +25,7 @@
     "@radix-ui/react-slot": "^1.0.2",
     "@tailwindcss/line-clamp": "^0.4.2",
     "@tanstack/react-query": "^4.29.7",
-    "axios": "^1.4.0",
+    "axios": "^1.6.2",
     "class-variance-authority": "^0.6.1",
     "clsx": "^1.2.1",
     "date-fns": "^2.30.0",
