+++ src/domain/projects/components/timers/modal/DeleteConfirmationModal.tsx
@@ -1,7 +1,7 @@
 import { Dialog, Transition } from '@headlessui/react';
-import { XMarkIcon, ExclamationTriangleIcon } from '@heroicons/react/24/outline';
+import { ExclamationTriangleIcon, XMarkIcon } from '@heroicons/react/24/outline';
 import React, { Fragment, ReactElement, useEffect } from 'react';
-import { EmptyTimer, TimerController, TimersListController } from '../hooks/types';
+import { EmptyTimer, TimerController, TimersListController } from '../../../types/timers';
 
 type Props = {
   timer_controller: TimerController;
